stability: address the comments
