--- tests/pkg/util/client/client.go
@@ -1,70 +1,0 @@
-// Copyright 2019 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-package client
-
-import (
-	"github.com/pingcap/errors"
-	"github.com/pingcap/tidb-operator/pkg/client/clientset/versioned"
-	"github.com/pingcap/tidb-operator/pkg/client/clientset/versioned/typed/pingcap.com/v1alpha1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
-)
-
-var (
-	masterUrl      string
-	kubeconfigPath string
-)
-
-type Client interface {
-	kubernetes.Interface
-	PingcapV1alpha1() v1alpha1.PingcapV1alpha1Interface
-}
-
-func Union(kube kubernetes.Interface, tidb versioned.Interface) Client {
-	return &client{Interface: kube, pingcap: tidb}
-}
-
-func NewOrDie() Client {
-	cfg, err := clientcmd.BuildConfigFromFlags(masterUrl, kubeconfigPath)
-	if err != nil {
-		panic(err)
-	}
-	return Union(kubernetes.NewForConfigOrDie(cfg), versioned.NewForConfigOrDie(cfg))
-}
-
-type client struct {
-	kubernetes.Interface
-	pingcap versioned.Interface
-}
-
-func (cli *client) PingcapV1alpha1() v1alpha1.PingcapV1alpha1Interface {
-	return cli.pingcap.PingcapV1alpha1()
-}
-
-func SetConfigPath(path string) {
-	kubeconfigPath = path
-}
-
-func SetMasterURL(url string) {
-	masterUrl = url
-}
-
-func LoadConfig() (*rest.Config, error) {
-	cfg, err := clientcmd.BuildConfigFromFlags(masterUrl, kubeconfigPath)
-	return cfg, errors.Trace(err)
-}
-
-type ClientOps struct {
-	Client
-}
